Rearrange Image implementations into multiple files
