+++ mirai-core/src/commonMain/kotlin/message/image/ImageUrlAware.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2019-2022 Mamoe Technologies and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
+ *
+ * https://github.com/mamoe/mirai/blob/dev/LICENSE
+ */
+
+package net.mamoe.mirai.internal.message.image
+
+import net.mamoe.mirai.Bot
+import net.mamoe.mirai.message.data.Image
+
+internal sealed interface ImageUrlAware : Image
+
+internal interface ConstOriginUrlAware : ImageUrlAware {
+    val originUrl: String
+}
+
+internal interface DeferredOriginUrlAware : ImageUrlAware {
+    fun getUrl(bot: Bot): String
+}
+
+internal interface SuspendDeferredOriginUrlAware : ImageUrlAware {
+    suspend fun getUrl(bot: Bot): String
+}
